UE Msg2 reception review
- enhancement of DL indication flow for Msg2 (RAR)
- merging interface procedures to handle DLSCH and RAR together
- better separation of MAC and PHY functionalities
- cleanup of double ambiguous functions to process RAR
- merging functions to process TA command into the ue_ta_procedures function
- enhancement of logging and formatting in the process

+++ openair1/PHY/defs_nr_UE.h
@@ -748,7 +748,7 @@ typedef struct NR_UL_TIME_ALIGNMENT {
   int16_t          ta_frame;
   char             ta_slot;
   /// TA command and TAGID received from the gNB
-  uint8_t          ta_command;
+  uint16_t         ta_command;
   uint8_t          tag_id;
 } NR_UL_TIME_ALIGNMENT_t;
